Finish some docstrings
Pylint's docparams plugin complains that the `:type`'s, `:rtype`'s and
`:raise` are missing from these docstrings.

Also changed EnvironmentError to ValueError. EnvironmentError seems to
confuse pylint's docparams checking and I think it might be because
EnvironmentError is only meant to be a base class for OSError and
others, not meant to be itself raised.