Don't use non-tox copy of Black
Because we're using sh to run black in tox.ini (rather than having tox
run black directly) sh will run whatever copy of black it finds first on
the user's $PATH. For example if the user has a system-wide copy of
black it will run that instead of tox's copy.

tox will not complain about a non-whitelisted executable from outside of
the virtualenv being executed either, because it's not tox that's
executing this non-virtualenv copy of black, it's sh that's executing
it.

We don't want to execute a copy of black from outside of the virtualenv
because this might not be the right version of black. Right now we don't
care what version of black is used and don't pin the version. But in the
future we might put a === or a >= on the black requirement in tox.ini
which would cause tox to ensure a certain version of black (for example
if a new version of black came out with a bug fix or new feature and we
wanted to auto-upgrade everyone's dev envs to it).

It also just introduces the possibility of the user having a broken or
somehow mis-installed global copy of black and that breaking their dev
env.

Force tox's version of black to be used by prefixing the black command
with tox's {envbindir} substitution (which is the path to the tox
venv's bin dir). See:

https://tox.readthedocs.io/en/latest/config.html#substitutions-for-virtualenv-related-sections