Refuse to `make release` if working tree unclean
Refuse to run `make release` if the git working tree appears to be
unclean.

This is because `make release` both builds the release archives (in
`dist/*`) from the contents of the working tree, and tags the git HEAD
commit with the version number.

If the contents of the HEAD commit and the working tree were different,
then what was tagged would not be the same contents as what was built.